Disabling opening the side menu
Sometimes it would be nice to be able to disable opening the side menu even while there is a view controller. I'm not entirely sure about the naming of the property.

My use case is that I have a subview of my navigation controller where I want to disable the side menu. I don't want to completely remove it just for this one view.